Refactor: extract sparql serialization from Draft
Extracts a SPARQL `ChangesetSerializer` from the `Draft` code. Draft now uses
this serializer by default.
