+++ lib/tufts/draft.rb
@@ -14,7 +14,9 @@ module Tufts
     #   @return [String]
     # @!attribute [rw] model
     #   @return [Tufts::Draftable]
-    attr_accessor :apply_strategy, :changeset, :id, :model
+    # @!attribute [rw] serializer
+    #   @return [Tufts::ChangesetSerializer]
+    attr_accessor :apply_strategy, :changeset, :id, :model, :serializer
 
     ##
     # @!method apply
@@ -49,6 +51,7 @@ module Tufts
       self.changeset      = changeset
       self.id             = id
       self.model          = model
+      self.serializer     = ChangesetSerializer.new
       self.apply_strategy = Tufts::ChangesetOverwriteStrategy
                             .new(model: model, changeset: changeset)
     end
@@ -83,9 +86,14 @@ module Tufts
     # Loads the changeset from the saved version.
     # @return [Draft] self
     def load
-      return @changeset = NullChangeSet.new unless exists?
+      if exists?
+        File.open(path) do |f|
+          self.changeset = serializer.deserialize(f.read, model: model)
+        end
+      else
+        @changeset = NullChangeSet.new
+      end
 
-      File.open(path) { |f| from_sparql(f.read) }
       self
     end
 
@@ -101,80 +109,14 @@ module Tufts
     #
     # @return [Draft] self
     def save
-      File.open(path, 'w+') { |f| f.write(to_sparql) }
+      File.open(path, 'w+') do |f|
+        f.write(serializer.serialize(changeset: changeset, subject: model.rdf_subject))
+      end
+
       self
     end
 
     private
-
-      ##
-      # We can't just apply the sparql string, because <> (null relative uri's)
-      # used by ActiveFedora is unsupported in SPARQL, so we need to parse and
-      # reconstruct the insert set manually.
-      #
-      # @return [void]
-      def from_sparql(update_string)
-        statements = extract_insert_statements(update_string)
-
-        self.changeset =
-          ActiveFedora::ChangeSet.new(model,
-                                      changed_graph(statements),
-                                      changed_attribute_keys(statements))
-        # cache the changes
-        changeset.changes
-      end
-
-      ##
-      # @return [RDF::Graph]
-      def changed_graph(statements)
-        graph = model.resource.dup
-
-        grouped = statements.group_by do |statement|
-          statement.subject.to_base + statement.predicate.to_base
-        end
-
-        grouped.each do |_, sts|
-          graph.update(sts.pop)
-          graph.insert(*sts)
-        end
-
-        graph
-      end
-
-      ##
-      # @return [Array<Symbol>]
-      def changed_attribute_keys(statements)
-        statements.map(&:predicate).uniq.map do |predicate|
-          config_for_predicate(predicate)
-        end.compact.map(&:first)
-      end
-
-      def config_for_predicate(predicate)
-        model.class.properties.find { |_, v| v.predicate == predicate }
-      end
-
-      ##
-      # @return [Array<RDF::Statement>)
-      def extract_insert_statements(update_string)
-        return [] if update_string.nil? || update_string.empty?
-
-        inserts = SPARQL.parse(update_string, update: true).each_descendant.select do |op|
-          op.is_a? SPARQL::Algebra::Operator::Insert
-        end
-
-        inserts.flat_map do |insert|
-          insert.operand.map do |pattern|
-            RDF::Statement(pattern.subject, pattern.predicate, pattern.object)
-          end
-        end
-      end
-
-      ##
-      # @return [String] a SPARQL Update string
-      def to_sparql
-        return '' if changeset.empty?
-        ActiveFedora::SparqlInsert.new(changeset.changes, model.rdf_subject).build
-      end
 
       def path
         STORAGE_DIR.join(id)
